Move SCC setup page into a separate module
