+++ .rubocop_todo.yml
@@ -1,15 +1,10 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2018-02-20 12:29:44 +0100 using RuboCop version 0.52.1.
+# on 2018-02-20 13:29:43 +0100 using RuboCop version 0.52.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
-
-# Offense count: 1
-Lint/UselessAssignment:
-  Exclude:
-    - 'src/lib/rmt/wizard.rb'
 
 # Offense count: 1
 Metrics/CyclomaticComplexity:
@@ -18,7 +13,7 @@ Metrics/CyclomaticComplexity:
 # Offense count: 2
 # Configuration parameters: CountComments.
 Metrics/MethodLength:
-  Max: 61
+  Max: 60
 
 # Offense count: 1
 Metrics/PerceivedComplexity:
@@ -27,5 +22,5 @@ Metrics/PerceivedComplexity:
 # Offense count: 2
 Style/MultipleComparison:
   Exclude:
-    - 'src/lib/rmt/wizard.rb'
     - 'src/lib/rmt/wizard_mariadb_page.rb'
+    - 'src/lib/rmt/wizard_scc_page.rb'
