+++ src/lib/rmt/wizard.rb
@@ -19,6 +19,7 @@
 require 'uri'
 require 'net/http'
 require 'rmt/base'
+require 'rmt/wizard_scc_page'
 require 'rmt/wizard_mariadb_page'
 
 module RMT
@@ -44,87 +45,13 @@ class RMT::Wizard < RMT::Base
   end
 
   def step1
-    contents = Frame(
-      _('SCC organization credentials'),
-      HBox(
-        HSpacing(1),
-        VBox(
-          VSpacing(1),
-          HSquash(
-            MinWidth(30, InputField(Id(:scc_username), _('Organization &username')))
-          ),
-          HSquash(
-            MinWidth(30, Password(Id(:scc_password), _('Organization &password')))
-          ),
-          VSpacing(1)
-        ),
-        HSpacing(1)
-      )
-    )
-
-    Wizard.SetContents(
-      _('RMT configuration step 1/2'),
-        contents,
-        "<p>Organization credentials can be found on Organization page at <a href='https://scc.suse.com/'>SUSE Customer Center</a>.</p>",
-        true,
-        true
-    )
-
-    Wizard.DisableBackButton
-
-    UI.ChangeWidget(Id(:scc_username), :Value, @config['scc']['username'])
-    UI.ChangeWidget(Id(:scc_password), :Value, @config['scc']['password'])
-
-    ret = nil
-    loop do
-      ret = UI.UserInput
-      if ret == :abort || ret == :cancel
-        break
-      elsif ret == :next
-        @config['scc']['username'] = Convert.to_string(UI.QueryWidget(Id(:scc_username), :Value))
-        @config['scc']['password'] = Convert.to_string(UI.QueryWidget(Id(:scc_password), :Value))
-
-        break if scc_credentials_valid?
-
-        break if Popup.AnyQuestion(
-          _('Invalid SCC credentials'),
-          _('SCC credentials are invalid. Please check the credentials.'),
-          _('Ignore and continue'),
-          _('Go back'),
-          :focus_no
-        )
-      end
-    end
-
-    deep_copy(ret)
+    page = RMT::WizardSCCPage.new(@config)
+    page.run
   end
 
   def step2
-    RMT::WizardMariaDBPage.new(@config)
-  end
-
-  def scc_credentials_valid?
-    UI.OpenDialog(
-      HBox(
-        HSpacing(5),
-        VBox(
-          VSpacing(5),
-          Left(Label(_('Checking SCC credentials...'))),
-          VSpacing(5)
-        ),
-        HSpacing(5)
-      )
-    )
-
-    uri = URI('https://scc.suse.com/connect/organizations/systems')
-    req = Net::HTTP::Get.new(uri)
-    req.basic_auth(@config['scc']['username'], @config['scc']['password'])
-
-    res = Net::HTTP.start(uri.host, uri.port, use_ssl: true) { |http| http.request(req) }
-
-    UI.CloseDialog
-
-    res.code.to_i == 200
+    page = RMT::WizardMariaDBPage.new(@config)
+    page.run
   end
 
   def run_wizard
@@ -141,8 +68,6 @@ class RMT::Wizard < RMT::Base
 
     Wizard.CreateDialog()
     Wizard.SetTitleIcon('yast-rmt')
-    Wizard.SetAbortButton(:abort, Label.CancelButton)
-    Wizard.SetNextButton(:next, Label.NextButton)
 
     Sequencer.Run(aliases, sequence)
